[MKT-2486] Cherry-pick commits to support images (#3)
* Images

The previous work in #124 by @ngscheurich and @joseustra
was using a twoCellAnchor but that's quite tricky
to get right.

This is now using a oneCellAnchor to make the
calculations much simpler.

The api is almost the same, but you don't need to add
any emu based values for offsets. You can also align
the image from the right of a cell. That's something
I needed but you also have to take the max character
width into account.

Also added a few extra tests for a few basic files.

I also fixed the formula because it didn't work
in the example file.

* Update some deps so it compiles

* Update warnings

---------

Co-authored-by: Kris Handley <handley.kris@gmail.com>